Use bigint for invoice.invoice_number

+++ server/repository/src/schema/number.rs
@@ -13,7 +13,7 @@ pub enum NumberRowType {
 #[table_name = "number"]
 pub struct NumberRow {
     pub id: String,
-    pub value: i32,
+    pub value: i64,
     /// Note, store id will be needed mainly for sync.
     pub store_id: String,
     // Table
